requirements.txt
pocketflow>=0.0.1
crawl4ai
google-cloud-aiplatform>=1.25.0
google-genai
openai>=1.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.2
requests>=2.31.0
playwright